refactor: rabbitmq 재시도 전략 설정으로 기존 코드 수정
