+++ server/src/main/java/com/vip/interviewpartner/domain/room_participant/service/ParticipantUpdateService.java
@@ -1,5 +1,6 @@
 package com.vip.interviewpartner.domain.room_participant.service;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.vip.interviewpartner.common.constants.Constants;
@@ -28,19 +29,14 @@ public class ParticipantUpdateService {
      * 참가자 관련 이벤트를 처리하는 리스너
      */
     @RabbitListener(queues = PARTICIPANT_EVENTS_QUEUE)
-    public void receiveParticipantEvent(String message) {
-        try {
-            JsonNode rootNode = objectMapper.readTree(message);
-            String eventType = rootNode.path("event").asText();
-            WebhookEvent webhookEvent = WebhookEvent.from(eventType);
-            switch (webhookEvent) {
-                case PARTICIPANT_JOINED -> handleParticipantJoined(rootNode);
-                case PARTICIPANT_LEFT -> handleParticipantLeft(rootNode);
-                default -> log.warn("Unhandled participant event type: {}", eventType);
-            }
-        } catch (Exception e) {
-            log.error("Failed to handle participant event", e);
-            // 필요 시, Dead Letter Queue로 메시지 전송하거나 재시도 로직 추가
+    public void receiveParticipantEvent(String message) throws JsonProcessingException {
+        JsonNode rootNode = objectMapper.readTree(message);
+        String eventType = rootNode.path("event").asText();
+        WebhookEvent webhookEvent = WebhookEvent.from(eventType);
+        switch (webhookEvent) {
+            case PARTICIPANT_JOINED -> handleParticipantJoined(rootNode);
+            case PARTICIPANT_LEFT -> handleParticipantLeft(rootNode);
+            default -> log.warn("Unhandled participant event type: {}", eventType);
         }
     }
